% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenDTau.R
\name{kenDTau}
\alias{kenDTau}
\title{kenDTau}
\usage{
kenDTau(v1, v2)
}
\arguments{
\item{v1}{The estimated topological sorting of a graph}

\item{v2}{The true topological sorting of a graph}
}
\value{
A value of Kendall’s tau
}
\description{
Provides Kendall’s tau to evaluate the estimated
     topological sorting of a graph
}
\examples{
v1 <- sample(1:10,10)
v2 <- 1:10
kenDTau(v1,v2)

}
