% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_NET.R
\name{PLOT_NET}
\alias{PLOT_NET}
\title{PLOT_NET}
\usage{
PLOT_NET(amatt)
}
\arguments{
\item{amatt}{An adjacent matrix of a graph}
}
\value{
A value of Kendall’s tau
}
\description{
Plot a netwok using an adjacent matrix of a graph
}
\examples{
gg <- PLOT_NET(res$amatt)
V(gg)$size = 18
plot(gg,layout=layout_with_kk(gg))
# the output of PLOT_NET can be operated as the igraph

}
