% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPearF.R
\name{SPearF}
\alias{SPearF}
\title{SPearF}
\usage{
SPearF(v1, v2)
}
\arguments{
\item{v1}{The estimated topological sorting of a graph}

\item{v2}{The true topological sorting of a graph}
}
\value{
A value of Spearman's footrule
}
\description{
Provides Spearman's footrule to evaluate the estimated
     topological sorting of a graph
}
\examples{
v1 <- sample(1:10,10)
v2 <- 1:10
SPearF(v1,v2)

}
